bot/requirements.txt
python-telegram-bot==20.3
websockets==11.0.3
python-dotenv==1.0.0
nest_asyncio